tools/necronomicon/requirements.txt
Babel==2.3.4
click==6.6
CoffeeScript==1.1.2
docutils==0.12
Flask==0.11.1
Flask-Mako==0.4
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
markdown2==2.3.1
MarkupSafe==0.23
Plim==0.9.12
PyExecJS==1.4.0
pyScss==1.3.5
pytz==2016.6.1
PyYAML==3.12
six==1.10.0
stylus==0.1.2
Werkzeug==0.11.11
